Removed old driver info files (see the ppd directory)
Documentation and const stuff in image library.

Added debug stuff to image library to aide in debugging problems that
crop up.

Added lseek() after seek() to ensure that the file pointer is updated
for TIFF files.

ImageRGBToCMYK() wasn't using the image density for the yellow plane.
This caused all sorts of problems...