The ESC/P decoder didn't report FF's in the print file.
Reverted back to the mink^3/maxk^2 formula for computing the black
channel, since it preserves more of the color in dark colors (no more
desaturation of colors...)

The EPSON drivers were not calling the Shutdown() function, so the
final reset sequence was not output.

The high-end EPSON drivers were not sending the "load defaults" remote
mode command after the job.

The true bottom margin for the SP 5500 is 50 points due to the location
of the paper and print head.  This caused the banding at the bottom of
each print...